sonyapi: adding awaitTakePicture to get imagesback from longexposure.
this is a stupid way to implement it. why does sony it not return the same way as when shooting from  cambody.
also when the cam gets set to idle and bevor awaitTakePicture was called the cam takes again a picture.
my workaround was to increase the timout to 15000ms.