Audio recording source set to high quality source
I've set it to CAMCORDER because it is camera audio source with stereo mic. 
DEFAULT uses the call mic that actively enables noise cancellation with  mono audio recording. 

I've previously posted a video on youtube because of a bug related to this in MI A1 OS https://www.youtube.com/watch?v=-DaW5b6mlcY.