try to get play upload working....
This release does not meet the 64-bit requirement of Google Play

The following APKs or app bundles are available on 64-bit devices, but only have native 32-bit code: 193.

You can add native 64- and 32-bit code to your app. With the Android App Bundle publication format, each device architecture receives only the native code that it needs. This will not unnecessarily increase the size of the app file.